requirements.txt
dungeon_despair_domain @ git+https://github.com/gallorob/dungeon-despair-domain
pygame_gui
pyyaml
pydantic
pillow<=11.1
fire
pyinstaller
requests